feat: add bash script to export content and push it to remote git repo
New variable to define in .env, env-sample: PAGES_REPO_PATH for where is the repo of the static page git based hosting.
Created a stage directory for where to put the above repo and make sure it is gitignored (as the content has its own repo).

Create bash script that first export the content using gssg in the public directory inside $PAGES_REPO_PATH,
then it will cd into that latter directory and git push to remote repo.

Refs: #4