refactor: move Dockerfile to gssg/Dockerfile as it's only for that dep.
The Dockerfile is not packaging the project but one of its dependency,
so it should not be at the project root. This is because convention says
that Dockerfile found at project root means the repo si for docker packageable tool.
This is an integration project, not a tool or command or service that can be shipped as
a container image.

Refs: #8