refactor: Use gssg from Docker Hub for immutability and security
to reduce exposure to supply-chain attacks on NPM, we freeze gssg and its dependencies
so we know they are not changing everytime this project is cloned or re-installed.

The first part was to build and push the image to Docker Hub after creating a namespace there.
The second part (this commit) is to update our docker-compose file to use the image from Docker hub.

Refs: #9