fix: prevent functions to use variables defined outside scope
why?

There is a scenario (forgetting to define a variable in a function, and have
that function be the first one called in the setup script)
where the automated tests are passing but the actual implementation in the script
fails.
This is because I use the same variable name everywhere, in all the functions
and by default bash variable seem to have global scope, so if I forget a definition,
it will pick up another definition outside scope.

how?

In all functions, we namespace all local variables to the function name
in setup script use a distinct variable name (and pass it to all functions that need to
know base directory for consistency)

Refs: #1