feat: add backup container service and backup bash script
How? initially planned to use bennetimo/ghost-backup:4 image, but it requires
hacking the ghost install to extract cookie file which I couldn't get working in order to save the JSON export.
There is an option to it to not perform JSON export backup, but even so it still tries to get cookies unnecessarily.
So I've forked bennetimo's repo and made a change to not try to create cookie file if the option to disable JSON backup
is disabled, hence the use of the image rija/ghost-backup:4 in the docker-compose.yml file.
the backup bash script needs to pass that option (-J) to the backup container

I've disabled the cron-based utomated backup (AUTOMATED_BACKUPS: false) because my setup is not running continuously.
I only fire the ghost server when I want to work on a blog post or publish one.
Backup is therefore done manually.

Refs: #5